market-backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
pymysql==1.1.0
mysqlclient==2.2.1
cryptography==42.0.5
twilio
dotenv
bcrypt
gunicorn==21.2.0
flask-cors